Adding morphological feature support
This is not a clean and generic implementation, but is tailored
for the Swedish model.
Note that we override CPOSTAG feature detectionand add it manually
as without it, the parser does not work (at least it does not produce
the same parse tree as the standalone version).
This implementation is not generic, hence not good for upstream.